+++ app/src/main/java/com/umc/sweepic/presentation/record/Adapter/ViewPagerAdapter.kt
@@ -2,9 +2,9 @@ package com.umc.sweepic.presentation.record.Adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.umc.sweepic.presentation.record.TagBoardFragment
 import com.umc.sweepic.presentation.record.history.HistoryFragment
 import com.umc.sweepic.presentation.record.memo.MemoFragment
+import com.umc.sweepic.presentation.record.tagboard.TagBoardFragment
 
 class ViewPagerAdapter(fragment : Fragment) : FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 3
