+++ app/src/main/java/com/umc/sweepic/presentation/record/RecordVpAdapter.kt
@@ -2,9 +2,9 @@ package com.umc.sweepic.presentation.record
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.umc.sweepic.presentation.record.History.HistoryFragment
-import com.umc.sweepic.presentation.record.Memo.MemoFragment
-import com.umc.sweepic.presentation.record.TagBord.TagBoardFragment
+import com.umc.sweepic.presentation.record.history.HistoryFragment
+import com.umc.sweepic.presentation.record.memo.MemoFragment
+import com.umc.sweepic.presentation.record.tagboard.TagBoardFragment
 
 class RecordVpAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 3
