+++ app/src/main/java/com/umc/sweepic/presentation/record/history/HistoryFragment.kt
@@ -1,42 +1,22 @@
 package com.umc.sweepic.presentation.record.history
 
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
+
 import com.umc.sweepic.R
 import com.umc.sweepic.databinding.FragmentHistoryBinding
-import com.umc.sweepic.presentation.record.history.HistoryFragment
+import com.umc.sweepic.presentation.base.BaseFragment
 
-
-class HistoryFragment : Fragment() {
-
-    private var _binding: FragmentHistoryBinding? = null
-    private val binding get() = _binding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentHistoryBinding.inflate(inflater, container, false)
-        return binding?.root
+class HistoryFragment : BaseFragment<FragmentHistoryBinding>(R.layout.fragment_history) {
+    override fun initObserver() {
+        // 필요하면 데이터 관찰 추가
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    override fun initView() {
+        // UI 초기화 추가
+    }
 
-        // ImageView 클릭 이벤트 추가
-        binding?.icNext?.setOnClickListener {
-            findNavController().navigate(R.id.historyTagFragment)
+    companion object {
+        fun newInstance(): HistoryFragment {
+            return HistoryFragment()
         }
     }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
 }
