--- app/src/main/res/navigation/challenge_graph.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    app:startDestination="@id/fragment_challenge">
-
-    <!-- Challenge Fragment -->
-    <fragment
-        android:id="@+id/fragment_challenge"
-        android:name="com.umc.sweepic.presentation.challenge.ChallengeFragment"
-        android:label="Challenge"
-        tools:layout="@layout/fragment_challenge">
-        <action
-            android:id="@+id/action_challenge_to_mypage"
-            app:destination="@id/fragment_mypage" />
-    </fragment>
-
-    <!-- MyPage Fragment -->
-    <fragment
-        android:id="@+id/fragment_mypage"
-        android:name="com.umc.sweepic.presentation.mypage.MyPageFragment"
-        android:label="MyPage"
-        tools:layout="@layout/fragment_mypage" />
-</navigation>
